automate required nav properties

+++ SM.DTOs/Dto/SlotDto.cs
@@ -5,7 +5,7 @@ namespace SM.Dto
     public abstract class SlotDto
     {
         public Guid Id { get; set; }
-        public int MinutesDuration { get; set; }
+        public byte MinutesDuration { get; set; }
         public byte Day { get; set; }
         public int Order { get; set; }
         public virtual ICollection<CourseTypeDto> CourseTypes { get; set; }
